Extracted interfaces from repositories classes, replaced usage of classes with interfaces, where possible.
