+++ src/main/java/com/andersen/repositories/BookRepository.java
@@ -3,44 +3,13 @@ package com.andersen.repositories;
 import com.andersen.enums.BookSortKey;
 import com.andersen.models.Book;
 
-import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
 
-public class BookRepository {
+public interface BookRepository {
+    List<Book> findAll();
 
-    private final List<Book> books;
+    Optional<Book> findByBookId(Long bookId);
 
-    public BookRepository(List<Book> books) {
-        this.books = books;
-    }
-
-
-    public List<Book> findAll() {
-        return books;
-    }
-
-    public Optional<Book> findByBookId(Long bookId) {
-        if (bookId < 1L) {
-            throw new IllegalArgumentException("Bad book id");
-        }
-        return books.stream()
-                .filter(book -> book.getId().equals(bookId))
-                .findFirst();
-    }
-
-    public List<Book> list(BookSortKey sortKey) {
-        List<Book> fetchedBooks = new ArrayList<>(books);
-        sort(fetchedBooks, sortKey);
-        return fetchedBooks;
-    }
-
-    public static void sort(List<Book> books, BookSortKey bookSortKey) {
-        switch (bookSortKey) {
-            case NAME -> books.sort(Comparator.comparing(Book::getName));
-            case PRICE -> books.sort(Comparator.comparing(Book::getPrice));
-            case AMOUNT -> books.sort(Comparator.comparing(Book::getAmount));
-        }
-    }
+    List<Book> list(BookSortKey sortKey);
 }
