+++ src/main/java/com/andersen/services/impl/RequestServiceImpl.java
@@ -3,6 +3,7 @@ package com.andersen.services.impl;
 import com.andersen.enums.RequestSortKey;
 import com.andersen.models.Request;
 import com.andersen.repositories.RequestRepository;
+import com.andersen.repositories.impl.RequestRepositoryDummy;
 import com.andersen.services.RequestService;
 
 import java.util.List;
